+++ lib/api/text_renderer.dart
@@ -2,9 +2,7 @@ import 'dart:math';
 
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'package:light_html_editor/api/parser.dart';
 import 'package:light_html_editor/api/richtext_node.dart';
-import 'package:light_html_editor/data/renderer_properties.dart';
 import 'package:light_html_editor/data/text_constants.dart';
 import 'package:light_html_editor/light_html_editor.dart';
 import 'package:url_launcher/url_launcher.dart';